add Vector::insert_ord_by[_key] methods
Users frequently want to use a custom ordering for their Vector types.
One way to accomplish this is to implement the Ord trait (along with
PartialOrd and usually PartialEq/Eq) using the desired custom logic for
comparisons.

However, this can lead to a lot of boilerplate code for what is often a
small, one-off use case. This has to to the Standard Library, and
libraries like imbl that offer similar APIs, creating methods such as
sort_by, sort_by_key, etc. that allow a user to pass a comparator
function that the library will use to sort or insert into the list.

This PR implements the insert_ord_by and insert_order_by_key methods,
providing a way to use a custom comparator function to insert into a
sorted list without needing to define the usual Rust comparison traits.