feat: Improve `Vector::skip`.
This patch improves `Vector::skip` by not calling `split_off` when the
`count` is equal to 0 or to `self.len()`.

Using `split_off` will allocate memory for everything before `count`.
This is a known problem.

`skip` continues to panic if `count > self.len()`. I'm not sure we want
to keep this behaviour since `Iterator::skip` doesn't panic in this
case, and I think it might be the reference here.